requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
asqlite==1.0.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
braceexpand==0.1.7
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
discord-ext-menus==1.0.0a32+g8686b5d
discord.py @ git+https://github.com/Rapptz/discord.py@c7e589033df300d0b305ca43ff7f1de361c9c4af
Flask==2.1.3
Flask-Threads==0.1.0
frozenlist==1.3.0
idna==3.3
import-expression==1.1.4
itsdangerous==2.1.2
Jinja2==3.1.2
jishaku==2.5.0
lark-parser==0.12.0
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.23.1
Pillow==9.2.0
six==1.16.0
Werkzeug==2.1.2
yarl==1.7.2
